refactor: 회원가입 API에 아이디 중복 확인 로직 추가
