+++ src/type/error/common-error.interface.ts
@@ -1,0 +1,8 @@
+export interface ICommonError {
+  status: number
+  code: string
+  message: {
+    [key: string]: string
+  }
+  origin?: string
+}
